+++ module/actor/actor-token.mjs
@@ -81,6 +81,15 @@ export class HeroSystem6eToken extends Token {
         return data;
     }
 
+    /**
+     * Add or remove the currently controlled Tokens from the active combat encounter
+     * @param {Combat} [combat]    A specific combat encounter to which this Token should be added
+     * @returns {Promise<Token>} The Token which initiated the toggle
+     */
+    async toggleCombat(combat) {
+        await super.toggleCombat(combat);
+    }
+
     _drawBar(number, bar, data) {
         // Ignore if bar3 not set
         if (!game.settings.get(game.system.id, "bar3")) {