+++ module/combat.mjs
@@ -1,6 +1,6 @@
 import { HEROSYS } from "./herosystem6e.mjs";
 import { clamp } from "./utility/compatibility.mjs";
-import { whisperUserTargetsForActor } from "./utility/util.mjs";
+import { whisperUserTargetsForActor, expireEffects } from "./utility/util.mjs";
 
 export class HeroSystem6eCombat extends Combat {
     constructor(data, context) {
@@ -12,8 +12,9 @@ export class HeroSystem6eCombat extends Combat {
     }
 
     getUniqueTokens() {
+        // this.combatants.contents.reduce((accumulator, item) => {if (!accumulator.find(o=> o.tokenId === item.tokenId)) {accumulator.push(item);} return accumulator;},[])
         const results = [];
-        for (const c of this.combatants.values()) {
+        for (const c of this.combatants.contents) {
             if (!results.find((o) => o.id === c.token.object?.id)) {
                 results.push(c.token.object);
             }
@@ -44,21 +45,76 @@ export class HeroSystem6eCombat extends Combat {
      */
 
     async rollInitiative(ids) {
+        // Structure input data
+        ids = typeof ids === "string" ? [ids] : ids;
+        if (ids.length === 0) {
+            ids = this.combatants.map((c) => c.id); // Roll All
+        }
+
         if (CONFIG.debug.combat) {
             console.debug(`Hero | rollInitiative`, ids);
         }
         // We need a unique combatant for each phase the actor acts.
 
-        for (const t of this.getUniqueTokens()) {
-            let tokenCombatants = this.combatants.filter(
-                (o) => o.tokenId === t.id && t.inCombat,
+        const uniqueTokensToProcess = [];
+
+        for (const id of ids) {
+            const c = this.combatants.find((o) => o.id === id);
+
+            const lightningReflexes = c.actor?.items.find(
+                (o) =>
+                    o.system.XMLID === "LIGHTNING_REFLEXES_ALL" ||
+                    o.system.XMLID === "LIGHTNING_REFLEXES_SINGLE",
             );
 
+            //Create extra combatants to match SPEED
+            let targetCombatantsForToken =
+                parseInt(
+                    Math.min(
+                        12,
+                        Math.max(1, c.actor?.system.characteristics.spd?.value), // Bases don't have a SPD
+                    ),
+                ) * (lightningReflexes ? 2 : 1);
+
+            const needToCreate =
+                targetCombatantsForToken -
+                this.combatants.filter((o) => o.tokenId === c.tokenId).length;
+
+            const toCreate = [];
+            for (let i = 0; i < needToCreate; i++) {
+                toCreate.push({
+                    tokenId: c.tokenId,
+                    sceneId: c.sceneId,
+                    actorId: c.actorId,
+                    hidden: c.hidden,
+                });
+            }
+            if (toCreate.length > 0) {
+                await this.createEmbeddedDocuments("Combatant", toCreate);
+                continue; // We just created combat documents, which will call this (RollInitiative) again
+            }
+
+            // Perhaps we have too many (SPD drain for example)
+            if (needToCreate < 0) {
+                await this.deleteEmbeddedDocuments("Combatant", [c.id], {
+                    single: true,
+                });
+                continue; // We just deleted combat documents, which will call this (RollInitiative) again
+            }
+
+            if (!uniqueTokensToProcess.find((t) => t.id === c.tokenId)) {
+                uniqueTokensToProcess.push(c.token);
+            }
+        }
+
+        // Loop thru all the tokens/combatants with ids provided
+        for (const t of uniqueTokensToProcess) {
             const lightningReflexes = t.actor?.items.find(
                 (o) =>
                     o.system.XMLID === "LIGHTNING_REFLEXES_ALL" ||
                     o.system.XMLID === "LIGHTNING_REFLEXES_SINGLE",
             );
+
             const lightningReflexesLevels = parseInt(
                 lightningReflexes?.system.LEVELS?.value ||
                     lightningReflexes?.system.LEVELS ||
@@ -67,51 +123,17 @@ export class HeroSystem6eCombat extends Combat {
                     0,
             );
 
-            //Create extra combatants to match SPEED
-            let needToCreate =
-                parseInt(
-                    Math.max(1, t.actor?.system.characteristics.spd.value),
-                ) *
-                    (lightningReflexes ? 2 : 1) -
-                tokenCombatants.length;
-
-            const toCreate = [];
-            for (let i = 0; i < needToCreate; i++) {
-                toCreate.push({
-                    tokenId: t.id,
-                    sceneId: t.scene.id,
-                    actorId: t.actor.id,
-                    hidden: t.document.hidden,
-                });
-            }
-
-            if (toCreate.length > 0) {
-                const created = await this.createEmbeddedDocuments(
-                    "Combatant",
-                    toCreate,
-                );
-                tokenCombatants = tokenCombatants.concat(created);
-            }
-
-            // Perhaps we have too many
-            const toDelete = [];
-            for (let i = 0; i < -needToCreate; i++) {
-                toDelete.push(tokenCombatants[i].id);
-            }
-            if (toDelete.length > 0) {
-                await this.deleteEmbeddedDocuments("Combatant", toDelete);
-            }
-
             // Produce an initiative roll for the Combatant.
             const characteristic =
-                tokenCombatants[0].actor?.system?.initiativeCharacteristic ||
-                "dex";
+                t.actor?.system?.initiativeCharacteristic || "dex";
             const initValue =
-                tokenCombatants[0].actor?.system.characteristics[characteristic]
-                    .value || 0;
-            const spdValue =
-                tokenCombatants[0].actor?.system.characteristics.spd.value || 0;
+                t.actor?.system.characteristics[characteristic]?.value || 0;
+            const spdValue = t.actor?.system.characteristics.spd?.value || 0;
             const initiativeValue = initValue + spdValue / 100;
+
+            const tokenCombatants = this.combatants.filter(
+                (c) => c.tokenId === t.id,
+            );
 
             // Assign a segment and Initiative
             let idx = 0;
@@ -120,8 +142,8 @@ export class HeroSystem6eCombat extends Combat {
                     HeroSystem6eCombat.hasPhase(
                         Math.max(
                             1,
-                            tokenCombatants[idx].actor?.system.characteristics
-                                .spd.value || 0,
+                            tokenCombatants[idx]?.actor?.system.characteristics
+                                .spd?.value || 0,
                         ),
                         s,
                     )
@@ -156,21 +178,26 @@ export class HeroSystem6eCombat extends Combat {
                         tokenCombatants[idx].flags.initiative !==
                             initiativeValue
                     ) {
-                        await tokenCombatants[idx].update({
-                            "flags.segment": s,
-                            initiative: initiativeValue,
-                            "-flags.lightningReflexesAlias": null,
-                        });
+                        try {
+                            await tokenCombatants[idx].update({
+                                "flags.segment": s,
+                                initiative: initiativeValue,
+                                "-flags.lightningReflexesAlias": null,
+                            });
+                        } catch (ex) {
+                            console.error(ex);
+                            return;
+                        }
                     }
                     idx++;
                 }
             }
 
-            // Rare case where SPD <= 0 (and for actorless tokens)
+            // Rare case where SPD <= 0 (and for actorless tokens; bases)
             // NOTE: There is no code to prevent a SPD 0 token from acting, currently GM player needs to handle that manually.
             // A SPD 0 character can't act, but does get a postSegment12.  In theory the SPD drain will eventually fade.
             if (
-                (tokenCombatants[0].actor?.system.characteristics.spd.value ||
+                (tokenCombatants[0].actor?.system.characteristics.spd?.value ||
                     0) <= 0 &&
                 tokenCombatants[0].flags.segment !== 12
             ) {
@@ -182,10 +209,6 @@ export class HeroSystem6eCombat extends Combat {
                 });
             }
         }
-
-        // Likely new turn orders so setupTurns
-        this.setupTurns();
-
         return this;
     }
 
@@ -326,11 +349,21 @@ export class HeroSystem6eCombat extends Combat {
         // Get current combatant
         const oldCombatant = this.combatant;
 
+        // Super
+        await super._onCreateDescendantDocuments(
+            parent,
+            collection,
+            documents,
+            data,
+            options,
+            userId,
+        );
+
         // Setup turns in segment fashion
         //const _turns = this.setupTurns();
 
         // Roll initiative (again?)
-        await this.rollInitiative();
+        await this.rollInitiative(documents.map((o) => o.id));
 
         // Keep the current Combatant the same after adding new Combatants to the Combat
         if (oldCombatant) {
@@ -341,16 +374,6 @@ export class HeroSystem6eCombat extends Combat {
             );
             await this.update({ turn: this.turn });
         }
-
-        // Super
-        await super._onCreateDescendantDocuments(
-            parent,
-            collection,
-            documents,
-            data,
-            options,
-            userId,
-        );
 
         // Render the collection
         //if (this.active) this.collection.render();
@@ -386,17 +409,20 @@ export class HeroSystem6eCombat extends Combat {
             userId,
         );
 
-        // Make sure we delete all combatants with the same tokenID
-        if (collection === "combatants") {
-            for (const doc of documents) {
-                const toDelete = this.combatants.filter(
-                    (c) => c.token.id === doc.token.id,
-                );
-                if (toDelete.length > 0) {
-                    await this.deleteEmbeddedDocuments(
-                        "Combatant",
-                        toDelete.map((c) => c.id),
+        // Make sure we delete all combatants with the same tokenID.
+        // Unless options.single = true; like when SPD lowers.
+        if (!options.single) {
+            if (collection === "combatants") {
+                for (const doc of documents) {
+                    const toDelete = this.combatants.filter(
+                        (c) => c.token.id === doc.token.id,
                     );
+                    if (toDelete.length > 0) {
+                        await this.deleteEmbeddedDocuments(
+                            "Combatant",
+                            toDelete.map((c) => c.id),
+                        );
+                    }
                 }
             }
         }
@@ -433,7 +459,7 @@ export class HeroSystem6eCombat extends Combat {
         }
         await this.update({ turn: this.turn, round: this.round });
 
-        await this.rollInitiative();
+        await this.rollInitiative(documents.map((o) => o.id));
 
         // Render the collection
         if (this.active) this.collection.render();
@@ -518,6 +544,10 @@ export class HeroSystem6eCombat extends Combat {
         await super._onStartTurn(combatant);
 
         if (!combatant) return;
+
+        // Expire Effects
+        // We expire on our phase, not on our segment.
+        await expireEffects(combatant.actor);
 
         // Reset movement history
         if (window.dragRuler) {
@@ -902,6 +932,9 @@ export class HeroSystem6eCombat extends Combat {
      * @returns {Promise<Combat>}
      */
     async nextTurn() {
+        if (CONFIG.debug.combat) {
+            console.debug(`Hero | nextTurn`);
+        }
         let turn = this.turn ?? -1;
         let skip = this.settings.skipDefeated;
 
@@ -928,13 +961,19 @@ export class HeroSystem6eCombat extends Combat {
         if (segment_next < segment) {
             segment_next += 12;
         }
-        let advanceTime = segment_next - segment;
+        const advanceTime = segment_next - segment;
 
         // Update the document, passing data through a hook first
         const updateData = { round, turn: next };
-        const updateOptions = { advanceTime: advanceTime, direction: 1 };
+        const updateOptions = {}; // { advanceTime: advanceTime, direction: 1 };
+
+        // Advance worldtime NOW, before we _onStartTurn gets called.
+        // Seems to be a bug in core foundryVTT.
+        await game.time.advance(advanceTime);
         Hooks.callAll("combatTurn", this, updateData, updateOptions);
-        return this.update(updateData, updateOptions);
+        let x = await this.update(updateData, updateOptions);
+        console.log(x);
+        return x;
     }
 
     /**